# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "ASE Project API"
  version: "1.0.0"
  description: "API documentation for ASE Project"
servers:
  - url: "http://localhost:5050"
paths:
...
  /api/admin/currency/{user_id}:
    get:
      summary: "Get transaction history for a user (Admin)"
      tags: ["AdminMicroservice"]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "ID of the user"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Transaction history retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '404':
          description: "User not found"
        '500':
          description: "Failed to retrieve transaction history"
...
  /currencytransaction:
    get:
      summary: "Get all currency transactions"
      tags: ["Currency Transaction"]
      responses:
        '200':
          description: "List of all currency transactions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CurrencyTransaction'
        '404':
          description: "Currency transactions not found"
    post:
      summary: "Create a new currency transaction"
      tags: ["Currency Transaction"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyTransactionRequest'
      responses:
        '200':
          description: "Currency transaction created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: integer
                    description: "ID of the newly created transaction"
        '400':
          description: "Invalid currency transaction data"
  /currencytransaction/{transactionId}:
    delete:
      summary: "Delete a currency transaction by ID"
      tags: ["Currency Transaction"]
      parameters:
        - in: path
          name: transactionId
          required: true
          schema:
            type: integer
          description: "ID of the transaction to delete"
      responses:
        '200':
          description: "Currency transaction deleted successfully"
        '404':
          description: "Currency transaction not found"
  /currencytransaction/{userId}:
    get:
      summary: "Get currency transactions for a specific user"
      tags: ["Currency Transaction"]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: "ID of the user"
      responses:
        '200':
          description: "List of currency transactions for the user"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CurrencyTransaction'
        '404':
          description: "Currency transactions not found"   
...
  /user/{userId}:
    get:
      summary: "Get a single user by ID"
      tags: ["User"]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: "ID of the user"
      responses:
        '200':
          description: "Details of the user"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: "User not found"
    put:
      summary: "Update a user by ID"
      tags: ["User"]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: "ID of the user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: "User updated successfully"
        '400':
          description: "Invalid user data"
    patch:
      summary: "Patch a user by ID"
      tags: ["User"]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: "ID of the user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: "User successfully updated"
        '400':
          description: "No data provided"
    delete:
      summary: "Delete a user by ID"
      tags: ["User"]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: "ID of the user"
      responses:
        '200':
          description: "User deleted successfully"
        '404':
          description: "User not found"
...
  /helloPlayer:
    get:
      summary: "Verify player token"
      tags: ["Player"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successfully verified player token"
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
        '500':
          description: "Internal server error"
  /helloAdmin:
    get:
      summary: "Verify admin token"
      tags: ["Admin"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successfully verified admin token"
          content:
            application/json:
              schema:
                type: object
                properties:
                  adminId:
                    type: integer
        '500':
          description: "Internal server error" 
...
components:  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          description: "Transaction ID"
        userId:
          type: integer
          description: "User ID for the transaction"
        amount:
          type: number
          format: float
          description: "Transaction amount"
        timestamp:
          type: string
          format: date-time
          description: "Transaction timestamp"
...
    CurrencyTransaction:
      type: object
      properties:
        id:
          type: integer
          description: "Unique ID of the transaction"
        userId:
          type: integer
          description: "ID of the user associated with the transaction"
        realAmount:
          type: number
          format: float
          description: "The real amount in currency"
        ingameAmount:
          type: number
          format: float
          description: "The ingame amount of currency"
        timestamp:
          type: string
          format: date-time
          description: "Timestamp of the transaction"
    CreateCurrencyTransactionRequest:
      type: object
      properties:
        userId:
          type: integer
          description: "ID of the user associated with the transaction"
        realAmount:
          type: number
          format: float
          description: "The real amount in currency"
        ingameAmount:
          type: number
          format: float
          description: "The ingame amount of currency"
        timestamp:
          type: string
          format: date-time
          description: "Timestamp of the transaction"  
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: "Unique ID of the user"
        authId:
          type: integer
          description: "Authentication ID of the user"
        profilePicture:
          type: string
          description: "URL of the user's profile picture"
        ingameCurrency:
          type: integer
          description: "Amount of in-game currency the user has"
        registrationDate:
          type: string
          format: date-time
          description: "Registration date of the user"
        status:
          type: string
          description: "Status of the user (e.g., ACTIVE, INACTIVE)"
        verified:
          type: boolean
          description: "Indicates whether the user is verified"
...
    UpdateUserRequest:
      type: object
      properties:
        ingameCurrency:
          type: integer
          description: "Updated in-game currency for the user"
        profilePicture:
          type: string
          description: "Updated profile picture URL"
        status:
          type: string
          description: "Updated status of the user"
...
security:
  - bearerAuth: []